backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.41
psycopg[binary]==3.2.10
pydantic-settings==2.11.0
apscheduler==3.11.1
python-dotenv==1.1.1
python-multipart==0.0.20
httpx==0.28.1
pytest==8.4.1
email-validator==2.3.0
